third_party/WebKit/LayoutTests/intersection-observer/containing-block-expected.txt

Test that no notifications are generated when root is not in the containing block chain of target.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 0
PASS entries.length is 1
PASS entries[0].boundingClientRect.left is 58
PASS entries[0].boundingClientRect.right is 158
PASS entries[0].boundingClientRect.top is 18
PASS entries[0].boundingClientRect.bottom is 118
PASS entries[0].intersectionRect.left is 58
PASS entries[0].intersectionRect.right is 158
PASS entries[0].intersectionRect.top is 18
PASS entries[0].intersectionRect.bottom is 118
PASS entries[0].rootBounds.left is 8
PASS entries[0].rootBounds.right is 193
PASS entries[0].rootBounds.top is 8
PASS entries[0].rootBounds.bottom is 208
PASS entries[0].target is [object HTMLDivElement]
PASS entries.length is 2
PASS entries[1].boundingClientRect.left is 58
PASS entries[1].boundingClientRect.right is 158
PASS entries[1].boundingClientRect.top is 258
PASS entries[1].boundingClientRect.bottom is 358
PASS entries[1].intersectionRect.left is 0
PASS entries[1].intersectionRect.right is 0
PASS entries[1].intersectionRect.top is 0
PASS entries[1].intersectionRect.bottom is 0
PASS entries[1].rootBounds.left is 8
PASS entries[1].rootBounds.right is 193
PASS entries[1].rootBounds.top is 8
PASS entries[1].rootBounds.bottom is 208
PASS entries[1].target is [object HTMLDivElement]
PASS entries.length is 2
PASS entries.length is 2
PASS successfullyParsed is true

TEST COMPLETE
